Add command for refreshing existing diagnostics from file and active onLanguage:csharp with initial diagnostic
