+++ src/constants.ts
@@ -1,1 +1,2 @@
 export const EXTENSION_NAME: string = "fakesharper";
+export const INSPECTION_FILENAME: string = "inspectcode.xml";
