+++ src/executor.ts
@@ -2,11 +2,12 @@ import { exec } from 'child_process';
 import { EOL } from 'os';
 import * as path from 'path';
 import * as vscode from 'vscode';
-import { EXTENSION_NAME } from './constants';
+import { EXTENSION_NAME, INSPECTION_FILENAME } from './constants';
 import { Issue } from './models';
 import { restorePath, getIssueSeverity, getIssueRange } from './utils';
 import { selectSolutionFile } from './workspace';
 import { parsefile } from './xmlparser';
+import { runDiagnostics } from './diagnostics';
 
 export class InspectCodeExecutor {
 	constructor(
@@ -34,60 +35,10 @@ export class InspectCodeExecutor {
 			} else {
 				const dirPath = path.dirname(filePath);
 
-				try {
-					const issues = parsefile(xmlPath);
-					restorePath(dirPath, issues);
-					this.updateDiagnostics(issues);
-				} catch (err) {
-					vscode.window.showErrorMessage(`${err?.message || err}`);
-				} finally {
-					this.hideStatusBarItem();
-				}
+				runDiagnostics(dirPath, this.diagnosticCollection);
+				this.hideStatusBarItem();
 			}
 		});
-	}
-
-	private updateDiagnostics(issues: Issue[]): void {
-		this.diagnosticCollection.clear();
-
-		type FileIssue = {
-			file: string;
-			issues: Issue[];
-		};
-
-		const fileIssues: FileIssue[] = [];
-
-		for (let i = 0; i < issues.length; i++) {
-			const issue: Issue = issues[i];
-			let fileIssueExists: boolean = false;
-			for (let j = 0; j < fileIssues.length; j++) {
-				const fileIssue: FileIssue = fileIssues[j];
-
-				if (issue.file === fileIssue.file) {
-					fileIssue.issues.push(issue);
-					fileIssueExists = true;
-					break;
-				}
-			}
-
-			if (!fileIssueExists) {
-				fileIssues.push({ file: issue.file, issues: [issue] });
-			}
-		}
-
-		for (let i = 0; i < fileIssues.length; i++) {
-			const fileIssue: FileIssue = fileIssues[i];
-
-			const uri: vscode.Uri = vscode.Uri.file(fileIssue.file);
-
-			this.diagnosticCollection.set(uri, fileIssue.issues.map(issue => ({
-				message: issue.message + (issue.issueType.wikiUrl ? EOL + issue.issueType.wikiUrl : ''),
-				range: getIssueRange(issue),
-				severity: getIssueSeverity(issue),
-				code: issue.typeId,
-				source: EXTENSION_NAME
-			})));
-		}
 	}
 
 	public run(): void {
@@ -97,7 +48,7 @@ export class InspectCodeExecutor {
 				return;
 			}
 
-			const xmlPath = path.join(path.dirname(filePath), 'inspectcode.xml');
+			const xmlPath = path.join(path.dirname(filePath), INSPECTION_FILENAME);
 
 			this.showStatusBarItem();
 			this.executeInspectCode(filePath, xmlPath);
