+++ src/extension.ts
@@ -1,12 +1,13 @@
 import * as vscode from 'vscode';
 import { EXTENSION_NAME } from './constants';
 import { CleanupCodeExecutor, InspectCodeExecutor } from './executor';
+import { runDiagnostics, runAllDiagnostics } from './diagnostics';
 
 export function activate(context: vscode.ExtensionContext) {
 	const diagnosticCollection = vscode.languages.createDiagnosticCollection(EXTENSION_NAME);
 
 	const statusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left);
-
+	
 	let disposable = vscode.commands.registerCommand(`${EXTENSION_NAME}.inspectcode`, () => {
 		new InspectCodeExecutor(statusBarItem, diagnosticCollection).run();
 	});
@@ -23,13 +24,20 @@ export function activate(context: vscode.ExtensionContext) {
 		new CleanupCodeExecutor(statusBarItem).run();
 	});
 
+	let disposable5 = vscode.commands.registerCommand(`${EXTENSION_NAME}.refreshdiagnostics`, () => {
+		runAllDiagnostics(statusBarItem, diagnosticCollection);
+	});
+	
 	context.subscriptions.push(
 		statusBarItem,
 		disposable,
 		disposable2,
 		disposable3,
-		disposable4
+		disposable4,
+		disposable5
 	);
+
+	runAllDiagnostics(statusBarItem, diagnosticCollection);
 }
 
 export function deactivate() { }
